Add files via upload
Heart disease is easier to treat when it is detected in the early stages. Machine learning
techniques may aid a more efficient analysis in the prediction of the disease. Moreover,
this prediction is one of the most central problems in medicine, as it is one of the leading
diseases related to an unhealthy lifestyle. So, an early prediction of this disease will be
useful for a cure or aversion. Here in this project we are predicting the chances having of heart disease
as well as analyzing the data to get some inferences using decision tree model.