Add files via upload
Abstract:
Marketing campaigns are characterized by focusing on the customer needs and their
overall satisfaction. Nevertheless, there are different variables that determine whether a
marketing campaign will be successful or not. There are certain variables that we need
to take into consideration when making a marketing campaign.
A Term deposit is a deposit that a bank or a financial institution offers with a fixed rate
(often better than just opening a deposit account) in which your money will be returned
back at a specific maturity time.
Problem Statement:
Predict if a customer subscribes to a term deposits or not, when contacted by a
marketing agent, by understanding the different features and performing predictive
analytics
DatasetInformation:
The data is related with direct marketing campaigns of a Portuguese banking
institution. The marketing campaigns were based on phone calls. Often, more than one
contact to the same client was required, in order to assess if the product (bank term
deposit) would be ('yes') or not ('no') subscribed.
The dataset consists of several predictor variables and one target variable, Outcome.
Predictor variables includes the age, job, marital status, and so on...
We have used Support Vector Machine Method to solve this problem and also class imbalance handling methods are used.